Update microsoft_model.py
For some users, the decoded token from the Microsoft access token does not have a key called 'email'  but a key called 'unique_name' that contains the user email address. Here, I propose to check if 'email' exists in the token_json, otherwise check for 'unique_name' to retrieve the user email address.